+++ metrics/coordination.go
@@ -1,0 +1,9 @@
+package metrics
+
+import (
+	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
+)
+
+func coordination(config Config) (t trace.Coordination) {
+	return t
+}