+++ metrics/retry.go
@@ -1,0 +1,10 @@
+package metrics
+
+import (
+	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
+)
+
+// Retry makes table.RetryTrace with New publishing
+func Retry(config Config) (t trace.Retry) {
+	return t
+}